feat: new dev/test environment
