+++ .dockerignore
@@ -1,0 +1,6 @@
+*
+!hcloud/
+!internal/
+!go.mod
+!go.sum
+!main.go
