+++ .gitignore
@@ -5,3 +5,5 @@ deploy/gen/
 .coverage.out
 .envrc
 hcloud-cloud-controller-manager
+*.tgz
+hack/.*
