--- e2etests/.gitignore
@@ -1,4 +1,0 @@
-ci-hcloud-ccm.tar
-ssh_key
-kubeconfig
-join.txt
