+++ hack/k3s-registries.yaml
@@ -1,0 +1,3 @@
+mirrors:
+  localhost:30666:
+    endpoint: ["http://10.43.0.2:5000"]
